Add user permission to context (#14)
* feat: add user permission management with caching support

* feat: implement user permission retrieval in JWT middleware and add mock user service tests

* feat: improve user service methods with cache invalidation and formatting adjustments

+++ sultan_core/tests/customer_repo.rs
@@ -765,7 +765,7 @@ async fn test_get_all_customers() {
             address: None,
             email: None,
             phone: None,
-            level: i as i32,
+            level: i,
             metadata: None,
         };
         repo.create(&ctx, id, &customer)
